chore: Split ActionViewDTO
